requirements.txt
langchain>=0.0.300
langgraph>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
tqdm>=4.64.0
